bugfix: search with empty
