+++ src/components/data-model-table/association-selection.tsx
@@ -21,6 +21,7 @@ import {
   setSelectedAssociatedItems,
 } from '../../redux/selectionSlice';
 import { LABEL_FIELD } from '../../util/decorators/LabelField';
+import { generateSearchFilters } from '../../util/tables';
 
 export interface AssociationSelectionProps<ParentModel, AssociatedModel>
   extends GqlAssociationProps {
@@ -205,12 +206,7 @@ export const AssociationSelection = <
 
   tableOptions.onSearch = useCallback(
     (values: any): CrudFilters => {
-      if (!values?.search?.length) return [];
-      return Array.from(searchableKeys).map((key) => ({
-        field: key,
-        operator: 'contains',
-        value: values.search,
-      }));
+      return generateSearchFilters(values, searchableKeys);
     },
     [searchableKeys],
   );
